v1.1 : Update realm val to support planes
- Added support to Run plane
- APIs to check if realm is running in P0 or Pn
- Added Plane related RMM structures, flags, enums

+++ test/exception/exception_realm_unsupported_smc/exception_realm_unsupported_smc_realm.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2023-2024, Arm Limited or its affiliates. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
@@ -7,14 +7,14 @@
 #include "exception_common_realm.h"
 void exception_realm_unsupported_smc_realm(void)
 {
-    int64_t retVal = 0;
+    uint64_t retVal = 0;
     uint8_t buff[100];
 
     val_set_status(RESULT_PASS(VAL_SUCCESS));
 
     /* smc call with invalid RSI/PSCI call */
-    retVal = (int64_t)(val_smc_call(RMI_REALM_CREATE, (uint64_t)buff,
-                                      0, 0, 0, 0, 0, 0, 0, 0, 0)).x0;
+    retVal = val_smc_call(RMI_REALM_CREATE, (uint64_t)buff,
+                                      0, 0, 0, 0, 0, 0, 0, 0, 0).x0;
 
     if (retVal == VAL_SMC_NOT_SUPPORTED)
     {
